requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pdfminer.six==20250506
groq==0.9.0
python-multipart==0.0.9
pydantic==2.11.9


